Separate NutClientMock class
